Seq refactoring
- Made Iterator the key lazy backing primitive
- Updated Concat

Feels like a rewrite of Seq with all of the new ref-struct ideas would be time well spent.  Just, later.

More long Counts

+++ LanguageExt.Core/Immutable Collections/Seq/Trait/Seq.TraitImpl.cs
@@ -125,7 +125,7 @@ public partial class Seq :
     static K<Seq, A> SemigroupK<Seq>.Combine<A>(K<Seq, A> ma, K<Seq, A> mb) =>
         ma.As() + mb.As();
 
-    static int Foldable<Seq>.Count<A>(K<Seq, A> ta) =>
+    static long Foldable<Seq>.Count<A>(K<Seq, A> ta) =>
         ta.As().Count;
 
     static FoldState IterableK<Seq, FoldState>.StepSetup<A>(K<Seq, A> ta) =>
